LockService Lock update change + transactional

+++ lock-microservice/service/src/main/java/io/github/safeslope/lock/service/LockService.java
@@ -2,11 +2,14 @@ package io.github.safeslope.lock.service;
 
 import io.github.safeslope.entities.Lock;
 import io.github.safeslope.lock.repository.LockRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+
 @Service
+@Transactional
 public class LockService {
     private final LockRepository repo;
 
@@ -32,7 +35,16 @@ public class LockService {
         return repo.save(lock);
     }
 
+    public Lock update(Integer id, Lock lock) {
+        repo.findById(id).orElseThrow(() -> new LockNotFoundException(id));
+        lock.setId(id);
+        return repo.save(lock);
+    }
+
     public void delete(Integer id) {
+        if (!repo.existsById(id)) {
+            throw new LockNotFoundException(id);
+        }
         repo.deleteById(id);
     }
 }
